+++ random.js
@@ -1,0 +1,5 @@
+function getRandom(num) {
+    return Math.ceil(Math.random() * num);
+}
+
+export default getRandom;
